Move rrdp and rsync back to the crate root.
