+++ src/lib.rs
@@ -37,7 +37,9 @@ pub mod operation;
 pub mod origins;
 pub mod output;
 pub mod repository;
+pub mod rrdp;
 pub mod rtr;
+pub mod rsync;
 pub mod slurm;
 pub mod utils;
 pub mod validity;
